feat(vue): added progress bar component and put range example in forms
