+++ bindings/vue/examples/components/Forms.vue
@@ -86,6 +86,23 @@
           {{ color }}
         </label>
       </ons-list-item>
+
+      <ons-list-header>Range</ons-list-header>
+      <ons-list-item>
+        Adjust the volume:
+        <ons-row>
+          <ons-col width="40px" style="text-align: center; line-height: 31px;">
+            <ons-icon icon="md-volume-down"></ons-icon>
+          </ons-col>
+          <ons-col>
+            <v-ons-range @input="onInput($event)" v-model.number="volume" style="width: 100%;"></v-ons-range>
+          </ons-col>
+          <ons-col width="40px" style="text-align: center; line-height: 31px;">
+            <ons-icon icon="md-volume-up"></ons-icon>
+          </ons-col>
+        </ons-row>
+        Volume: {{ volume }} <span id="loud-alert" style="display: none">(careful, that's loud)</span>
+      </ons-list-item>
     </ons-list>
   </v-ons-page>
 </template>
@@ -99,13 +116,23 @@
         vegetables: ['Tomato', 'Cabbage', 'Cucumber'],
         selectedVegetable: 'Cabbage',
         colors: ['Red', 'Blue', 'Yellow', 'Green'],
-        checkedColors: []
+        checkedColors: [],
+        volume: 50
       };
     },
 
     methods: {
       log: function(event) {
         console.log('qweqwe', event);
+      },
+      onInput(volume) {
+        const elem = document.getElementById("loud-alert");
+        if (volume > 80) {
+          elem.style.display = "inline-block";
+        }
+        else {
+          elem.style.display = "none";
+        }
       }
     }
 	};
