+++ bindings/vue/examples/components/index.js
@@ -7,8 +7,6 @@ import SpeedDial from './SpeedDial.vue';
 import Dialogs from './Dialogs.vue';
 import Carousel from './Carousel.vue';
 import Forms from './Forms.vue';
-import Range from './Range.vue';
-import ProgressBar from './ProgressBar.vue';
 
 export {
   Tabbar,
@@ -19,7 +17,5 @@ export {
   SpeedDial,
   Dialogs,
   Carousel,
-  Forms,
-  Range,
-  ProgressBar
+  Forms
 };
