+++ bindings/vue/src/components/index.js
@@ -10,7 +10,6 @@ export { default as VOnsSplitter } from './VOnsSplitter.vue';
 export { default as VOnsSplitterSide } from './VOnsSplitterSide.vue';
 export { default as VOnsSplitterContent } from './VOnsSplitterContent.vue';
 export { default as VOnsRange } from './VOnsRange.vue';
-export { default as VOnsProgressBar } from './VOnsProgressBar.vue';
 
 import VGeneric from './VGeneric.vue';
 const extend = component => Object.assign({name: 'v-ons-' + component}, VGeneric);
@@ -35,4 +34,5 @@ export const VOnsRipple = extend('ripple');
 export const VOnsRow = extend('row');
 export const VOnsCol = extend('col');
 export const VOnsModal = extend('modal');
+export const VOnsProgressBar = extend('progress-bar');
